+++ viscord-1.18.2-fabric-forge-template/common/src/main/java/network/vonix/viscord/discord/WebhookClient.java
@@ -32,8 +32,8 @@ public class WebhookClient {
     public void updateUrl(String webhookUrl) {
         this.webhookUrl = webhookUrl;
     }
-
-    public String getWebhookUrl() {
+    
+    public String getUrl() {
         return webhookUrl;
     }